<refacto> move widgets files to package views/widgets
