+++ Mobile/Carto/lib/views/widgets/map_widget.dart
@@ -1,13 +1,12 @@
-import 'package:carto/utils/buttons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:latlong2/latlong.dart';
 
-import 'package:carto/views/pages/search_page.dart';
 import '../services/location_service.dart';
 import '../../models/establishment.dart';
 import '../../views/services/establishment_service.dart';
+import 'buttons.dart';
 
 class MapWidget extends StatefulWidget {
   static final GlobalKey<_MapWidgetState> mapKey = GlobalKey<_MapWidgetState>();
@@ -172,37 +171,16 @@ class _MapWidgetState extends State<MapWidget> {
             ),
           ),
           Positioned(
-            bottom: 400,
-            right: 0,
-            child: FloatingActionButton(
-              onPressed: () {
-                Navigator.pushNamed(
-                  context,
-                  '/signup',
-                );
-              },
-              child: const Icon(Icons.account_circle_rounded),
-            ),
-          ),
-          Positioned(
             top: 70,
             left: 30,
-            child: BlueSquareIconButton(icon: Icons.search, onPressed: () {
-              Navigator.pushNamed(
-                context,
-                '/search',
-              );
-            }
-
-            /*FloatingActionButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => SearchPage()),
-                );
-              },
-              child: const Icon(Icons.search),*/
-            ),
+            child: BlueSquareIconButton(
+                icon: Icons.search,
+                onPressed: () {
+                  Navigator.pushNamed(
+                    context,
+                    '/search',
+                  );
+                }),
           ),
         ],
       ),
